requirements.txt
djangorestframework==3.15.1
flake8==7.0.0
psycopg2-binary==2.9.9